move beta_distribution to its own file
